+++ qsu/src/main/scala/quasar/qsu/mra/Provenance.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quasar.qscript.provenance
+package quasar.qsu.mra
 
 import slamdata.Predef.{Boolean, Vector}
