+++ qsu/src/test/scala/quasar/qsu/ExpandShiftsSpec.scala
@@ -16,6 +16,8 @@
 
 package quasar.qsu
 
+import slamdata.Predef.{Map => _, _}
+
 import quasar.{Qspec, TreeMatchers}
 import quasar.IdStatus.{ExcludeId, IdOnly}
 import quasar.contrib.iota._
@@ -23,11 +25,9 @@ import quasar.contrib.matryoshka._
 import quasar.ejson.{EJson, Fixed}
 import quasar.ejson.implicits._
 import quasar.fp._
-import quasar.impl.provenance.ProvImpl
 import quasar.qscript.{construction, Hole, OnUndefined, PlannerError, SrcHole}
 import quasar.qsu.{QScriptUniform => QSU}
-import slamdata.Predef.{Map => _, _}
-import quasar.contrib.iota.{copkEqual, copkTraverse}
+import quasar.qsu.mra.ProvImpl
 
 import matryoshka._
 import matryoshka.data._
