+++ qsu/src/test/scala/quasar/qsu/mra/CatsNonEmptyListGenerator.scala
@@ -14,25 +14,32 @@
  * limitations under the License.
  */
 
-package quasar.qscript.provenance
+package quasar.qsu.mra
 
 import slamdata.Predef.List
 
 import quasar.pkg.tests._
 
-import cats.Eq
+import scala.Predef.$conforms
+
+import cats.data.NonEmptyList
 
 import org.scalacheck.Cogen
 
-trait UopGenerator {
-  implicit def arbitraryUop[A: Arbitrary: Eq]: Arbitrary[Uop[A]] =
+// Copied from cats-laws due to the scalacheck version incompatibility between
+// cats-1.5.0 and quasar, using the instances from `cats` results in
+// NoSuchMethodErrors at runtime.
+//
+// FIXME: Remove once cats upgrades to scalacheck-1.14.x
+trait CatsNonEmptyListGenerator {
+  implicit def catsArbitraryNel[A: Arbitrary]: Arbitrary[NonEmptyList[A]] =
     Arbitrary(for {
-      sz <- Gen.frequency((32, 1), (16, 2), (8, 3), (4, 4), (2, 5), (1, 0))
-      as <- Gen.listOfN(sz, arbitrary[A])
-    } yield Uop.of(as: _*))
+      h <- arbitrary[A]
+      t <- arbitrary[List[A]]
+    } yield NonEmptyList(h, t))
 
-  implicit def cogenUop[A: Cogen]: Cogen[Uop[A]] =
+  implicit def catsCogenNel[A: Cogen]: Cogen[NonEmptyList[A]] =
     Cogen[List[A]].contramap(_.toList)
 }
 
-object UopGenerator extends UopGenerator
+object CatsNonEmptyListGenerator extends CatsNonEmptyListGenerator
