+++ qsu/src/test/scala/quasar/qsu/mra/JoinKeyGenerator.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package quasar.qscript.provenance
+package quasar.qsu.mra
 
 import quasar.pkg.tests._
